+++ libs/movie/src/lib/movie/pages/summary/promotional/media-videos/media-videos.component.scss
@@ -1,0 +1,8 @@
+strong {
+  opacity: 0.5;
+  font-weight: normal;
+}
+
+h6 {
+  margin: 0;
+}
