+++ libs/movie/src/lib/movie/pages/summary/summary.component.html
@@ -107,120 +107,15 @@
     <section id="promotional-elements">
       <h3>Promotional Elements</h3>
       <article>
-        <h6>
-          Files
-          <a mat-icon-button [routerLink]="getPath('promotional-elements')" fragment="files"><mat-icon svgIcon="pencil" color="primary"></mat-icon></a>
-        </h6>
-        <ul>
-          <li>
-            <mat-icon svgIcon="eye" color="primary"></mat-icon><strong>Presentation Deck </strong>
-            <ng-container *ngIf="(form.promotional.get('presentation_deck').value.ref | isEmptyImgRef) else presentationDeck">
-              <a [routerLink]="getPath('promotional-elements')" fragment="files" class="missing-control">
-                Missing
-              </a>
-            </ng-container>
-            <ng-template #presentationDeck>
-              {{ form.promotional.get('presentation_deck').value.fileName }}
-            </ng-template>
-          </li>
-          <li>
-            <mat-icon svgIcon="eye" color="primary"></mat-icon><strong>Scenario </strong>
-            <ng-container *ngIf="(form.promotional.get('scenario').value.ref | isEmptyImgRef) else scenario">
-              <a [routerLink]="getPath('promotional-elements')" fragment="files" class="missing-control">
-                Missing
-              </a>
-            </ng-container>
-            <ng-template #scenario>
-              {{ form.promotional.get('scenario').value.fileName }}
-            </ng-template>
-          </li>
-          <li>
-            <mat-icon svgIcon="eye" color="primary"></mat-icon><strong>Presentation Deck </strong>
-            <ng-container *ngIf="(form.promotional.get('moodboard').value.ref | isEmptyImgRef) else moodboard">
-              <a [routerLink]="getPath('promotional-elements')" fragment="files" class="missing-control">
-                Missing
-              </a>
-            </ng-container>
-            <ng-template #moodboard>
-              {{ form.promotional.get('moodboard').value.fileName }}
-            </ng-template>
-          </li>
-        </ul>
+        <movie-summary-media-files movie="form" [link]="getPath('media-files')"></movie-summary-media-files>
       </article>
 
       <article>
-        <h6>
-          Images
-          <a mat-icon-button [routerLink]="getPath('media-images')"><mat-icon svgIcon="pencil" color="primary"></mat-icon></a>
-        </h6>
-        <ul>
-          <ng-container *ngIf="objectHasNoValue(stillPhoto) else displayImages">
-            <a [routerLink]="getPath('media-images')" class="missing-control">
-              Missing
-            </a>
-          </ng-container>
-          <ng-template #displayImages>
-            <li *ngFor="let image of stillPhoto.controls | keyvalue; let i = index; let isLast = last">
-              <mat-icon svgIcon="eye" color="primary"></mat-icon><strong>Image {{ i+1 }} </strong>
-              <a [href]="getPath('media-images')">
-                {{ image.value.oldRef.value | fileName }}
-              </a>
-            </li>
-          </ng-template>
-        </ul>
+        <movie-summary-media-images movie="form" [link]="getPath('media-images')"></movie-summary-media-images>
       </article>
 
       <article>
-        <h6>
-          Videos
-          <a mat-icon-button [routerLink]="getPath('promotional-elements')" fragment="videos"><mat-icon svgIcon="pencil" color="primary"></mat-icon></a>
-        </h6>
-        <ul>
-          <li>
-            <mat-icon svgIcon="eye" color="primary"></mat-icon><strong>Teaser Link </strong>
-            <ng-container *ngIf="!form.promotional.get('teaser_link').value else teaserLink">
-              <a [routerLink]="getPath('promotional-elements')" fragment="videos" class="missing-control">
-                Missing
-              </a>
-            </ng-container>
-            <ng-template #teaserLink>
-              {{ form.promotional.get('teaser_link').value }}
-            </ng-template>
-          </li>
-          <li>
-            <mat-icon svgIcon="eye" color="primary"></mat-icon><strong>Trailer Link </strong>
-            <ng-container *ngIf="!form.promotional.get('trailer_link').value else trailerLink">
-              <a [routerLink]="getPath('promotional-elements')" fragment="videos" class="missing-control">
-                Missing
-              </a>
-            </ng-container>
-            <ng-template #trailerLink>
-              {{ form.promotional.get('trailer_link').value }}
-            </ng-template>
-          </li>
-          <li>
-            <mat-icon svgIcon="eye" color="primary"></mat-icon><strong>Promo Link </strong>
-            <ng-container *ngIf="!form.promotional.get('promo_reel_link').value else promoLink">
-              <a [routerLink]="getPath('promotional-elements')" fragment="videos" class="missing-control">
-                Missing
-              </a>
-            </ng-container>
-            <ng-template #promoLink>
-              {{ form.promotional.get('promo_reel_link').value }}
-            </ng-template>
-          </li>
-          <li>
-            <mat-icon svgIcon="eye" color="primary"></mat-icon><strong>Screener Link </strong>
-            <ng-container *ngIf="!form.promotional.get('screener_link').value else screenerLink">
-              <a [routerLink]="getPath('promotional-elements')" fragment="files" class="missing-control">
-                Missing
-              </a>
-            </ng-container>
-            <ng-template #screenerLink>
-              {{ form.promotional.get('screener_link').value }}
-            </ng-template>
-          </li>
-        </ul>
+        <movie-summary-media-videos movie="form" [link]="getPath('media-videos')"></movie-summary-media-videos>
       </article>
 
     </section >
